refactor: 타임테이블에 추가할 페스티벌 목록 조회 쿼리옵션 이동
