+++ apps/client/src/shared/apis/performance/performance-queries.ts
@@ -1,4 +1,4 @@
-import { infiniteQueryOptions, queryOptions } from '@tanstack/react-query';
+import { queryOptions } from '@tanstack/react-query';
 
 import { get } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
@@ -6,7 +6,6 @@ import { PERFORMANCE_QUERY_KEY } from '@shared/constants/query-key';
 import { BaseResponse } from '@shared/types/api';
 import { ConcertDetailResponse } from '@shared/types/concert-response';
 import { FestivalDetailResponse } from '@shared/types/festival-response';
-import { GetFestivalToAddResponse } from '@shared/types/get-festival-to-add-response';
 
 export const PERFORMANCE_QUERY_OPTIONS = {
   CONCERT: (concertId: number) =>
@@ -18,13 +17,6 @@ export const PERFORMANCE_QUERY_OPTIONS = {
     queryOptions({
       queryKey: PERFORMANCE_QUERY_KEY.FESTIVAL(festivalId),
       queryFn: () => getFestivalDetail(festivalId),
-    }),
-  GET_FESTIVAL_TO_ADD_LIST: () =>
-    infiniteQueryOptions<GetFestivalToAddResponse, Error>({
-      queryKey: PERFORMANCE_QUERY_KEY.GET_FESTIVAL_TO_ADD.LIST(),
-      queryFn: ({ pageParam }) => getFestivalToAdd(pageParam as number),
-      initialPageParam: undefined,
-      getNextPageParam: (lastPage) => lastPage.nextCursor || undefined,
     }),
 };
 
@@ -42,15 +34,6 @@ export const getFestivalDetail = async (
 ): Promise<FestivalDetailResponse> => {
   const response = await get<BaseResponse<FestivalDetailResponse>>(
     `${END_POINT.GET_FESTIVAL_DETAIL}/${festivalId}`,
-  );
-  return response.data;
-};
-
-export const getFestivalToAdd = async (
-  cursor?: number,
-): Promise<GetFestivalToAddResponse> => {
-  const response = await get<BaseResponse<GetFestivalToAddResponse>>(
-    END_POINT.GET_FESTIVAL_TO_ADD(cursor),
   );
   return response.data;
 };
